Simple C++ Program.
Simple C++ Program
In C++, any line starting with // is a comment.It is completely ignored by the C++ compiler.
The #include is a preprocessor directive used to include files in our program.
We need to use #include <iostream> to use cout that allows us to print output on the screen.
int main() {...} _ A valid C++ program must have the main() function. The curly braces indicate the start and the end of the function.
The execution of code beings from this function.
cout prints the inside the quotation marks. It must be followed by << followed by the format string. 
In our example, "Welcome to github" is the format string.
Semicolon(;) is used to indicate the end of a statement.
The return 0; - statement is the "Exit status" of the program. In simple terms, the program ends with this statement.